+++ BlazarData/src/main/java/com/hubspot/blazar/data/service/BranchService.java
@@ -30,7 +30,7 @@ public class BranchService {
       return gitInfo;
     } else {
       try {
-        long id = branchDao.insert(gitInfo);
+        int id = branchDao.insert(gitInfo);
         return gitInfo.withId(id);
       } catch (UnableToExecuteStatementException e) {
         if (e.getCause() instanceof SQLIntegrityConstraintViolationException) {
